Move tests next to the source to allow linting tests.
