+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-    roots: ["<rootDir>/test"],
+    roots: ["<rootDir>/src"],
     testMatch: ["**/__tests__/**/*.+(ts|tsx|js)", "**/?(*.)+(spec|test).+(ts|tsx|js)"],
     transform: {
         "^.+\\.(ts|tsx)$": "ts-jest",
